refactoring: improve separation between interface and implementation
Earlier, AbstractRewardMachine was defined as a subclass of pythomata.DFA.
This introduced a strong coupling with the pythomata package.
We now decoupled this by adding new abstract methods/properties to the interface:

- states
- initial_state
- get_transitions_from
- get_transitions

The method get_successor now mandatorily requires to return a state.

The concrete class that implements this interface is now called RewardAutomaton.

We also have a separation between AbstractRewardMachineSimulator and
RewardMachine simulator; the former is just an interface, whereas the latter
is a generic RM simulator that takes in input an *AbstractRewardMachine*
instance, and *not* a concrete class of AbstractRewardMachine. This will
make RewardMachineSimulator work with any AbstractRewardMachine instance.

Another important change is that now TemporalGoal does not take in input
the 'reward' value. The class TemporalGoal is still useful as a composition
of both RM and RMSimulator.